bots: Move import_module_from_source to finder.py.
